+++ compilation_maker/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.12.4"
+__version__ = "0.12.5"
 APP_NAME = "CompilationMaker"
 
 # When you make a user-facing change to Memoir Montage, bump __version__ above.